move discovery types and admission types into kube-core
