+++ CHANGELOG.md
@@ -4,7 +4,7 @@ UNRELEASED
  * see https://github.com/clux/kube-rs/compare/0.53.0...master
  * `kube`: `api` `discovery` module now uses a new `ApiResource` struct #495
  * `kube`: `api` BREAKING: `DynamicObject` now takes an `ApiResource` rather than a `GroupVersionKind`
- * `kube`: `api` `discovery` module's `Group` renamed to `ApiGroup`
+ * `kube`: `api` BREAKING: `discovery` module's `Group` renamed to `ApiGroup`
  * `kube-core` crate factored out of `kube` to reduce dependencies - #516 via #517
 
 0.53.0 / 2021-05-15
