+++ kube-core/Cargo.toml
@@ -13,6 +13,7 @@ readme = "../README.md"
 
 [features]
 ws = []
+admission = ["json-patch"]
 
 [dependencies]
 serde = { version = "1.0.118", features = ["derive"] }
@@ -22,6 +23,7 @@ once_cell = "1.7.2"
 url = "2.2.0"
 log = "0.4.11"
 http = "0.2.2"
+json-patch = { version = "0.2.6", optional = true }
 
 [dependencies.k8s-openapi]
 version = "0.11.0"
