+++ kube/src/api/mod.rs
@@ -1,17 +1,7 @@
 //! API helpers for structured interaction with the Kubernetes API
 
-use serde::{Deserialize, Serialize};
 
-/// Empty struct for when data should be discarded
-///
-/// Not using [`()`](https://doc.rust-lang.org/stable/std/primitive.unit.html), because serde's
-/// [`Deserialize`](serde::Deserialize) `impl` is too strict.
-#[derive(Clone, Deserialize, Serialize, Default, Debug)]
-pub struct NotUsed {}
-
-pub(crate) mod typed;
-pub use typed::Api;
-
+mod core_methods;
 #[cfg(feature = "ws")] mod remote_command;
 #[cfg(feature = "ws")] pub use remote_command::AttachedProcess;
 
@@ -21,16 +11,107 @@ pub use subresource::{AttachParams, Attachable, Executable};
 pub use subresource::{EvictParams, Evictable, LogParams, Loggable, ScaleSpec, ScaleStatus};
 
 // Re-exports from kube-core
+#[cfg(feature = "admission")] pub use kube_core::admission;
 pub(crate) use kube_core::params;
 pub use kube_core::{
     dynamic::{ApiResource, DynamicObject},
     gvk::{GroupVersionKind, GroupVersionResource},
     metadata::{ListMeta, ObjectMeta, Resource, ResourceExt, TypeMeta},
-    object::{Object, ObjectList, WatchEvent},
+    object::{NotUsed, Object, ObjectList, WatchEvent},
     request::Request,
 };
 pub use params::{
     DeleteParams, ListParams, Patch, PatchParams, PostParams, Preconditions, PropagationPolicy,
 };
 
-#[cfg(feature = "admission")] pub mod admission;
+use crate::Client;
+/// The generic Api abstraction
+///
+/// This abstracts over a [`Request`] and a type `K` so that
+/// we get automatic serialization/deserialization on the api calls
+/// implemented by the dynamic [`Resource`].
+#[derive(Clone)]
+pub struct Api<K> {
+    /// The request builder object with its resource dependent url
+    pub(crate) request: Request,
+    /// The client to use (from this library)
+    pub(crate) client: Client,
+    /// Note: Using `iter::Empty` over `PhantomData`, because we never actually keep any
+    /// `K` objects, so `Empty` better models our constraints (in particular, `Empty<K>`
+    /// is `Send`, even if `K` may not be).
+    pub(crate) phantom: std::iter::Empty<K>,
+}
+
+/// Api constructors for Resource implementors with custom DynamicTypes
+///
+/// This generally means resources created via [`DynamicObject`](crate::api::DynamicObject).
+impl<K: Resource> Api<K> {
+    /// Cluster level resources, or resources viewed across all namespaces
+    ///
+    /// This function accepts `K::DynamicType` so it can be used with dynamic resources.
+    pub fn all_with(client: Client, dyntype: &K::DynamicType) -> Self {
+        let url = K::url_path(dyntype, None);
+        Self {
+            client,
+            request: Request::new(url),
+            phantom: std::iter::empty(),
+        }
+    }
+
+    /// Namespaced resource within a given namespace
+    ///
+    /// This function accepts `K::DynamicType` so it can be used with dynamic resources.
+    pub fn namespaced_with(client: Client, ns: &str, dyntype: &K::DynamicType) -> Self {
+        let url = K::url_path(dyntype, Some(ns));
+        Self {
+            client,
+            request: Request::new(url),
+            phantom: std::iter::empty(),
+        }
+    }
+
+    /// Consume self and return the [`Client`]
+    pub fn into_client(self) -> Client {
+        self.into()
+    }
+
+    /// Return a reference to the current resource url path
+    pub fn resource_url(&self) -> &str {
+        &self.request.url_path
+    }
+}
+
+
+/// Api constructors for Resource implementors with Default DynamicTypes
+///
+/// This generally means structs implementing `k8s_openapi::Resource`.
+impl<K: Resource> Api<K>
+where
+    <K as Resource>::DynamicType: Default,
+{
+    /// Cluster level resources, or resources viewed across all namespaces
+    pub fn all(client: Client) -> Self {
+        let url = K::url_path(&Default::default(), None);
+        Self {
+            client,
+            request: Request::new(url),
+            phantom: std::iter::empty(),
+        }
+    }
+
+    /// Namespaced resource within a given namespace
+    pub fn namespaced(client: Client, ns: &str) -> Self {
+        let url = K::url_path(&Default::default(), Some(ns));
+        Self {
+            client,
+            request: Request::new(url),
+            phantom: std::iter::empty(),
+        }
+    }
+}
+
+impl<K> From<Api<K>> for Client {
+    fn from(api: Api<K>) -> Self {
+        api.client
+    }
+}
